requirements.txt
seaborn
scikit-learn
matplotlib
numpy
pandas
torch
torchvision
torchaudio
huggingface
transformers
streamlit
yaspin
python-dotenv
wikipedia-api